+++ Controller/Subjects.cs
@@ -2,7 +2,7 @@
 using System.Collections;
 using System.Collections.Generic;
 
-namespace University_advisor.Controllers
+namespace University_advisor.Controller
 {
     public class Subjects : IEnumerable
     {
