+++ Entity/Review.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 
-namespace University_advisor.Controllers
+namespace University_advisor.Controller
 {
     public class Review
     {
