+++ Uni adv.csproj
@@ -65,8 +65,8 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Controllers\Serializer.cs" />
-    <Compile Include="Controllers\Review.cs" />
+    <Compile Include="Controller\Serializer.cs" />
+    <Compile Include="Entity\Review.cs" />
     <Compile Include="View\ConfirmationBox.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -85,9 +85,9 @@
     <Compile Include="View\Menu.Designer.cs">
       <DependentUpon>Menu.cs</DependentUpon>
     </Compile>
-    <Compile Include="Controllers\Subjects.cs" />
-    <Compile Include="Controllers\SubjectEnum.cs" />
-    <Compile Include="Controllers\Subject.cs" />
+    <Compile Include="Controller\Subjects.cs" />
+    <Compile Include="Controller\SubjectEnum.cs" />
+    <Compile Include="Entity\Subject.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <EmbeddedResource Include="View\ConfirmationBox.resx">
