+++ View/Menu.cs
@@ -3,7 +3,7 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using System.Windows.Forms;
-using University_advisor.Controllers;
+using University_advisor.Controller;
 
 namespace University_advisor
 {
