+++ View/SubjectCard.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using System.Windows.Forms;
-using University_advisor.Controllers;
+using University_advisor.Controller;
 using University_advisor.View;
 
 namespace University_advisor
